Deprecate duplicate and copyToClipboard icon alias
After material icons upgrade, the former filled copy/duplicate icon
is not filled anymore and looks exactly the same as the same but outlined
copyToClipboard icon, which means that the outlined icon
now represents all three aliases. We agreed to only use `copy`
alias for the icon which is the most commonly alias used
in our products.