KDropdownMenu emits close and tab events for fine-tuned focus management
When a user hits Tab they will leave the focus of the popover but currently the focus just returns to the root of the page rather than the next item in the dom.

Here we do two things - first we emit a "close" event so that users can react on close. This is less necessary to the stated purpose but seems a worthy addition to allow us to react to the closing of a dropdown.

The second is to emit a `tab` event which occurs when the user hits `tab`. Note that in this case, we pass the JS event so that it can be `preventDefault`-ed if needed (which was the motivation for this to begin with)

lint