Move a11y related validators to BaseCard
It's BaseCard that is utilizing heading level
and title, and therefore that's the place to ensure
that these data are provided in the expected format.

This also removes pieces of validation there weren't
needed. For example, Vue takes care of checking
non-empty value for required props, so we don't
need to do it.