[WIP] KCard can have a checkbox
To be able to complete this in a11y way,
KCheckbox will need to be updated to allow
for more flexible label. Due to how KCheckbox
is implemented it seems that currently there
is no way how to set the card title as its checkbox
label.

There may be some more a11y requirements needed,
to be discussed.