Fix screenreaders announcing all card content
when navigating the list of cards with TAB key.

This resolves overwhelming output and aligns the screenreader
navigation to the expected experience when on TAB only card titles
should be announced (other content is still accessible by  navigating
with arrows).

The main culprit that caused all content of cards being announced
was 'tabindex=0' set on <li>. To resolve the problem,
an alternative technique that doesn't require this tabindex is used
to make the focus ring display around the whole card area.