Fix incorrect position of label
When inner slots were present, the label
was centered, but it should be on the left
side. Centered position was likely use to fix
overlapping of the label with the content
of the left inner slot in an empty input.
This is now resolved by ensuring that the label
is positioned in the upper area of the input
(=not applying transform style) when an inner slot
is present.